+++ servers/src/main/ru/itmo/mit/nonblockingserver/NonBlockingServer.java
@@ -107,12 +107,12 @@ public class NonBlockingServer implements Server {
     }
 
     @Override
-    public long getRequestProcessingTime() {
+    public int getRequestProcessingTime() {
         throw new UnsupportedOperationException();
     }
 
     @Override
-    public long getClientProcessingTime() {
+    public int getClientProcessingTime() {
         throw new UnsupportedOperationException();
     }
